Update des Client Sourcecodes.
git-svn-id: https://invertika.googlecode.com/svn/trunk@1304 784bf6b6-0765-11df-962c-b5c252e50527

+++ src/net/manaserv/playerhandler.h
@@ -1,8 +1,9 @@
 /*
- *  The Mana World
- *  Copyright (C) 2004-2010  The Mana World Development Team
+ *  The Mana Client
+ *  Copyright (C) 2004-2009  The Mana World Development Team
+ *  Copyright (C) 2009-2010  The Mana Developers
  *
- *  This file is part of The Mana World.
+ *  This file is part of The Mana Client.
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -15,8 +16,7 @@
  *  GNU General Public License for more details.
  *
  *  You should have received a copy of the GNU General Public License
- *  along with this program; if not, write to the Free Software
- *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 #ifndef NET_MANASERV_PLAYERHANDLER_H
@@ -44,31 +44,23 @@ class PlayerHandler : public MessageHandler, public Net::PlayerHandler
         void handleMessage(Net::MessageIn &msg);
 
         void attack(int id);
-
         void emote(int emoteId);
 
-        void increaseAttribute(size_t attr);
-
-        void decreaseAttribute(size_t attr);
-
+        void increaseAttribute(int attr);
+        void decreaseAttribute(int attr);
         void increaseSkill(int skillId);
 
         void pickUp(FloorItem *floorItem);
-
         void setDirection(char direction);
-
         void setDestination(int x, int y, int direction = -1);
-
         void changeAction(Being::Action action);
 
         void respawn();
 
         void ignorePlayer(const std::string &player, bool ignore);
-
         void ignoreAll(bool ignore);
 
         bool canUseMagic();
-
         bool canCorrectAttributes();
 
         int getJobLocation();
